requirements.txt
psycopg2-binary
requests
python-dotenv
numpy
pandas
sqlalchemy